only remove posts in queue from queue

+++ html/js/admin.js
@@ -226,21 +226,25 @@ function loadAdminUpvotesTable(upvotes) {
 
     // Delete
     newcolumn = document.createElement('td');
-    newlink = document.createElement('a');
-    newlink.setAttribute('href','#');
-    newlink.setAttribute('id','deleteUpvote'+value.id);
-    newimage = document.createElement('img');
-    newimage.setAttribute('src','img/icons/trash.svg');
-    newimage.setAttribute('height','24px');
-    newlink.appendChild(newimage);
-    newcolumn.appendChild(newlink);
+    if(value.status == 'in queue') {
+      newlink = document.createElement('a');
+      newlink.setAttribute('href','#');
+      newlink.setAttribute('id','deleteUpvote'+value.id);
+      newimage = document.createElement('img');
+      newimage.setAttribute('src','img/icons/trash.svg');
+      newimage.setAttribute('height','24px');
+      newlink.appendChild(newimage);
+      newcolumn.appendChild(newlink);
+    }
     newrow.appendChild(newcolumn);
 
     document.getElementById('adminUpvotesTableBody').appendChild(newrow);
 
-    document.getElementById("deleteUpvote"+value.id).onclick = function() {
-      if(confirm('Delete upvote for '+value.title+'?') == true) {
-        deleteUpvote(value.id);
+    if(value.status == 'in queue') {
+      document.getElementById("deleteUpvote"+value.id).onclick = function() {
+        if(confirm('Delete upvote for '+value.title+'?') == true) {
+          deleteUpvote(value.id);
+        }
       }
     }
   });
